# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "E-Commerce API Documentation"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "tags": ["System"],
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string", "example": "Electronics"},
                  "description": {"type": "string", "example": "Electronic products"},
                  "parent_id": {"type": "integer", "example": null},
                  "is_active": {"type": "boolean", "example": true}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category created successfully"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get a specific category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully"
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "description": {"type": "string"},
                  "is_active": {"type": "boolean"},
                  "parent_id": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully"
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "example": 1}
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          }
        }
      }
    }
  }
}